# Week_1/Day_3_and_4/java_server/src/main/java/org/openapitools/api/BooksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.20.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Book;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-22T18:23:51.347862200+05:30[Asia/Calcutta]", comments = "Generator version: 7.20.0")
@Validated
@Tag(name = "books", description = "the books API")
public interface BooksApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_BOOKS_BOOK_ID_DELETE = "/books/{bookId}";
    /**
     * DELETE /books/{bookId} : Delete a book by its ID
     *
     * @param bookId  (required)
     * @return Book deleted successfully (status code 204)
     *         or Book not found (status code 404)
     */
    @Operation(
        operationId = "booksBookIdDelete",
        summary = "Delete a book by its ID",
        responses = {
            @ApiResponse(responseCode = "204", description = "Book deleted successfully"),
            @ApiResponse(responseCode = "404", description = "Book not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = BooksApi.PATH_BOOKS_BOOK_ID_DELETE,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> booksBookIdDelete(
        @NotNull @Parameter(name = "bookId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("bookId") Integer bookId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKS_BOOK_ID_GET = "/books/{bookId}";
    /**
     * GET /books/{bookId} : Find a book by its ID
     *
     * @param bookId  (required)
     * @return Book is found (status code 200)
     *         or Book is not found (status code 404)
     */
    @Operation(
        operationId = "booksBookIdGet",
        summary = "Find a book by its ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Book is found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            }),
            @ApiResponse(responseCode = "404", description = "Book is not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_BOOKS_BOOK_ID_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<Book> booksBookIdGet(
        @NotNull @Parameter(name = "bookId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("bookId") Integer bookId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"author\" : \"Mayukh Haldar\", \"available\" : true, \"id\" : 100, \"title\" : \"A Book\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKS_BOOK_ID_PUT = "/books/{bookId}";
    /**
     * PUT /books/{bookId} : Update a book by its ID
     *
     * @param bookId  (required)
     * @param book  (required)
     * @return Book Updated successfully (status code 200)
     *         or Book not found (status code 404)
     */
    @Operation(
        operationId = "booksBookIdPut",
        summary = "Update a book by its ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Book Updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            }),
            @ApiResponse(responseCode = "404", description = "Book not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = BooksApi.PATH_BOOKS_BOOK_ID_PUT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Book> booksBookIdPut(
        @NotNull @Parameter(name = "bookId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("bookId") Integer bookId,
        @Parameter(name = "Book", description = "", required = true) @Valid @RequestBody Book book
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"author\" : \"Mayukh Haldar\", \"available\" : true, \"id\" : 100, \"title\" : \"A Book\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKS_GET = "/books";
    /**
     * GET /books : Get a list of all books
     *
     * @return A successful list of books (status code 200)
     */
    @Operation(
        operationId = "booksGet",
        summary = "Get a list of all books",
        responses = {
            @ApiResponse(responseCode = "200", description = "A successful list of books", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Book.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_BOOKS_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Book>> booksGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"author\" : \"Mayukh Haldar\", \"available\" : true, \"id\" : 100, \"title\" : \"A Book\" }, { \"author\" : \"Mayukh Haldar\", \"available\" : true, \"id\" : 100, \"title\" : \"A Book\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKS_POST = "/books";
    /**
     * POST /books : Add a new book
     *
     * @param book  (required)
     * @return Book added successfully (status code 201)
     *         or Invalid input (status code 400)
     */
    @Operation(
        operationId = "booksPost",
        summary = "Add a new book",
        responses = {
            @ApiResponse(responseCode = "201", description = "Book added successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = BooksApi.PATH_BOOKS_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Book> booksPost(
        @Parameter(name = "Book", description = "", required = true) @Valid @RequestBody Book book
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"author\" : \"Mayukh Haldar\", \"available\" : true, \"id\" : 100, \"title\" : \"A Book\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}