# Week_1/Day_3_and_4/java_server/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: An API to manage Books
  title: The Bookstore API
  version: 1.0.0
servers:
- url: http://localhost:8080/api/v1
paths:
  /books:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Book"
                type: array
          description: A successful list of books
      summary: Get a list of all books
      x-accepts:
      - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
          description: Book added successfully
        "400":
          content:
            application/json:
              schema:
                type: object
          description: Invalid input
      summary: Add a new book
      x-content-type: application/json
      x-accepts:
      - application/json
  /books/{bookId}:
    delete:
      parameters:
      - explode: false
        in: path
        name: bookId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          content: {}
          description: Book deleted successfully
        "404":
          content:
            application/json:
              schema:
                type: object
          description: Book not found
      summary: Delete a book by its ID
      x-accepts:
      - application/json
    get:
      parameters:
      - explode: false
        in: path
        name: bookId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
          description: Book is found
        "404":
          content:
            application/json:
              schema:
                type: object
          description: Book is not found
      summary: Find a book by its ID
      x-accepts:
      - application/json
    put:
      parameters:
      - explode: false
        in: path
        name: bookId
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
          description: Book Updated successfully
        "404":
          content:
            application/json:
              schema:
                type: object
          description: Book not found
      summary: Update a book by its ID
      x-content-type: application/json
      x-accepts:
      - application/json
components:
  schemas:
    Book:
      example:
        author: Mayukh Haldar
        available: true
        id: 100
        title: A Book
      properties:
        id:
          example: 100
          type: integer
        title:
          example: A Book
          type: string
        author:
          example: Mayukh Haldar
          type: string
        available:
          example: true
          type: boolean
      required:
      - author
      - title
      type: object